requirements.txt
py-cord==2.0.0b5
Pillow
requests
sqlalchemy